Moved the supportedLanguages to a dedicated file Added a test to verify that all the files are imported as selectedLanguages
